Move ddtrace_dogstatsd_url function in a new file as dogstatsd_client.c is not compiled on windows
