+++ ext/dogstatsd.h
@@ -1,0 +1,6 @@
+#ifndef DDTRACE_DOGSTATSD_H
+#define DDTRACE_DOGSTATSD_H
+
+char *ddtrace_dogstatsd_url(void);
+
+#endif  // DDTRACE_DOGSTATSD_H
