+++ ext/dogstatsd_client.h
@@ -6,6 +6,5 @@
 void ddtrace_dogstatsd_client_minit(void);
 void ddtrace_dogstatsd_client_rinit(void);
 void ddtrace_dogstatsd_client_rshutdown(void);
-char *ddtrace_dogstatsd_url(void);
 
 #endif  // DDTRACE_DOGSTATSD_CLIENT_H
